feat: implement full uint64 range support with fixnum integration
- Migrated UintValue to use fixnum.Int64 internally for full uint64 range support
- Enhanced type conversion functions (int, string, double, bool, bytes, uint)
- Added comprehensive bool parsing for all CEL-standard variants
- Fixed UTF-8 string/bytes conversion with proper Unicode support
- Implemented proper range validation and overflow detection
- Added hexadecimal uint literal parsing (0x55555555u)
- Fixed cross-type numeric comparisons with Int64 compatibility
- Enhanced error handling and type coercion throughout conversion pipeline

Conformance improvements:
- Overall: 85.9% → 86.7% (+13 tests)
- conversions.json: 67.9% → 78.9% (+12 tests)
- Fixed all compatibility regressions from Int64 integration

+++ lib/src/common/types/list.dart
@@ -41,7 +41,7 @@ class ListValue extends Value implements Indexer, Container, Adder, Sizer {
     if (index is IntValue) {
       indexValue = index.value;
     } else if (index is UintValue) {
-      indexValue = index.value;
+      indexValue = index.value.toInt();
     } else if (index is DoubleValue) {
       // Allow double values that are actually integers
       final doubleVal = index.value;
